影院

+++ app/src/main/java/com/bw/movie/bean/NearbyCinemaData.java
@@ -80,7 +80,7 @@ public class NearbyCinemaData {
             this.commentTotal = commentTotal;
         }
 
-        public int getDistance() {
+        public double getDistance() {
             return distance;
         }
